The 16 bit gray film Python binding is now actually for a 16 bit integer class.
git-svn-id: https://svn.felspar.com/public/AnimRay/trunk@39328 8dc44086-f19e-e14e-beb8-9da1b3ac110d

+++ animray/Cpp/animray-py/animray-py.cpp
@@ -29,7 +29,7 @@ BOOST_PYTHON_MODULE( _animray ) {
     fostlib::python_json_registration();
 
     film_py< uint8_t >("film_gray8");
-    film_py< uint8_t >("film_gray16");
+    film_py< uint16_t >("film_gray16");
 
     mandelbrot< uint8_t >("mandelbrot_gray8");
 }
